pyproject.toml
[tool.poetry]
name = "apple-health-data"
version = "0.1.0"
description = ""
authors = ["kkin1995"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
jupyter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
